# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AgencyOS API
  description: Backend for the AI Outreach Agency Operating System.
  version: 0.1.0
paths:
...
  /api/v1/monitoring/operational/summary:
    get:
      tags:
      - monitoring
      summary: Get operational summary dashboard
      description: 'Get comprehensive operational summary for production dashboards.


        Returns execution counts, worker health, automation status, schedule stats,

        and system health metrics.'
      operationId: get_operational_summary_api_v1_monitoring_operational_summary_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationalSummaryResponse'
      security:
      - HTTPBearer: []
  /api/v1/monitoring/execution-statistics:
    get:
      tags:
      - monitoring
      summary: Get execution statistics
      description: 'Get detailed execution statistics for the specified time window.


        Returns counts by execution status, workflow distribution, and organization

        distribution.'
      operationId: get_execution_statistics_api_v1_monitoring_execution_statistics_get
      security:
      - HTTPBearer: []
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Time window in hours
          default: 24
          title: Hours
        description: Time window in hours
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionStatisticsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/worker-statistics:
    get:
      tags:
      - monitoring
      summary: Get worker statistics
      description: 'Get worker health and activity statistics.


        Returns worker counts, health status, error distribution, and detailed worker

        information.'
      operationId: get_worker_statistics_api_v1_monitoring_worker_statistics_get
      security:
      - HTTPBearer: []
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Time window in hours
          default: 24
          title: Hours
        description: Time window in hours
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkerStatisticsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/schedule-statistics:
    get:
      tags:
      - monitoring
      summary: Get schedule statistics
      description: 'Get schedule dispatch statistics.


        Returns schedule dispatcher metrics including queued, failed, skipped, and

        conflict counts.'
      operationId: get_schedule_statistics_api_v1_monitoring_schedule_statistics_get
      security:
      - HTTPBearer: []
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Time window in hours
          default: 24
          title: Hours
        description: Time window in hours
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ScheduleStatisticsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/retention-statistics:
    get:
      tags:
      - monitoring
      summary: Get retention statistics
      description: 'Get retention worker statistics.


        Returns retention sweep metrics including deleted execution events and

        pruned worker health records.'
      operationId: get_retention_statistics_api_v1_monitoring_retention_statistics_get
      security:
      - HTTPBearer: []
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Time window in hours
          default: 24
          title: Hours
        description: Time window in hours
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetentionStatisticsResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/automation-lifecycle:
    get:
      tags:
      - monitoring
      summary: Get automation lifecycle statistics
      description: 'Get automation lifecycle statistics including pause/resume events.


        Returns counts of automation pause/resume events and current automation

        status from the control service.'
      operationId: get_automation_lifecycle_api_v1_monitoring_automation_lifecycle_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AutomationLifecycleResponse'
      security:
      - HTTPBearer: []
  /api/v1/monitoring/heartbeat-visibility:
    get:
      tags:
      - monitoring
      summary: Get worker heartbeat visibility
      description: 'Get worker heartbeat visibility for monitoring.


        Returns list of worker instances with their heartbeat status, useful for

        identifying stale or failed workers.'
      operationId: get_heartbeat_visibility_api_v1_monitoring_heartbeat_visibility_get
      security:
      - HTTPBearer: []
      parameters:
      - name: worker_type
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by worker type (e.g., 'execution')
          title: Worker Type
        description: Filter by worker type (e.g., 'execution')
      - name: stale_within_seconds
        in: query
        required: false
        schema:
          type: integer
          maximum: 86400
          minimum: 60
          description: Staleness window in seconds
          default: 300
          title: Stale Within Seconds
        description: Staleness window in seconds
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 1000
          minimum: 1
          description: Maximum results
          default: 100
          title: Limit
        description: Maximum results
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/HeartbeatVisibilityResponse'
                title: Response Get Heartbeat Visibility Api V1 Monitoring Heartbeat Visibility Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/execution-timeline:
    get:
      tags:
      - monitoring
      summary: Get execution timeline events
      description: 'Get recent execution timeline events across all organizations.


        Returns the latest execution events with workflow name and execution

        status/duration attached, optionally filtered by status or workflow.'
      operationId: get_execution_timeline_api_v1_monitoring_execution_timeline_get
      security:
      - HTTPBearer: []
      parameters:
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Time window in hours
          default: 24
          title: Hours
        description: Time window in hours
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ExecutionStatus'
          - type: 'null'
          description: Filter by execution status
          title: Status
        description: Filter by execution status
      - name: workflow
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by workflow name (partial match)
          title: Workflow
        description: Filter by workflow name (partial match)
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          description: Maximum results
          default: 100
          title: Limit
        description: Maximum results
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionTimelineResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/execution-history:
    get:
      tags:
      - monitoring
      summary: Get execution history
      description: Get paginated execution history across all organizations.
      operationId: get_execution_history_api_v1_monitoring_execution_history_get
      security:
      - HTTPBearer: []
      parameters:
      - name: page
        in: query
        required: false
        schema:
          type: integer
          minimum: 1
          description: Page number (1-based)
          default: 1
          title: Page
        description: Page number (1-based)
      - name: page_size
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          description: Results per page
          default: 50
          title: Page Size
        description: Results per page
      - name: hours
        in: query
        required: false
        schema:
          type: integer
          maximum: 168
          minimum: 1
          description: Time window in hours
          default: 24
          title: Hours
        description: Time window in hours
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/ExecutionStatus'
          - type: 'null'
          description: Filter by execution status
          title: Status
        description: Filter by execution status
      - name: workflow
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by workflow name (partial match)
          title: Workflow
        description: Filter by workflow name (partial match)
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExecutionHistoryResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/monitoring/queue-status:
    get:
      tags:
      - monitoring
      summary: Get queue status
      description: Get per-organization queue status across all organizations.
      operationId: get_queue_status_api_v1_monitoring_queue_status_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QueueStatusResponse'
      security:
      - HTTPBearer: []
  /api/v1/monitoring/monitoring-information:
    get:
      tags:
      - monitoring
      summary: Get comprehensive monitoring information
      description: Get comprehensive system, database, worker, and queue information.
      operationId: get_monitoring_information_api_v1_monitoring_monitoring_information_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MonitoringInformationResponse'
      security:
      - HTTPBearer: []
...
components:
  schemas:
...
    AutomationLifecycleResponse:
      properties:
        automation_paused_events:
          type: integer
          title: Automation Paused Events
        automation_resumed_events:
          type: integer
          title: Automation Resumed Events
        current_status:
          additionalProperties: true
          type: object
          title: Current Status
      type: object
      required:
      - automation_paused_events
      - automation_resumed_events
      - current_status
      title: AutomationLifecycleResponse
      description: Automation lifecycle statistics (pause/resume events).
...
    DatabaseHealth:
      properties:
        connected:
          type: boolean
          title: Connected
        pool_usage:
          type: number
          title: Pool Usage
        active_connections:
          type: integer
          title: Active Connections
        avg_query_time:
          type: number
          title: Avg Query Time
      type: object
      required:
      - connected
      - pool_usage
      - active_connections
      - avg_query_time
      title: DatabaseHealth
      description: Database connection and pool health.
...
    ExecutionHistoryEntry:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        workflow_id:
          type: string
          format: uuid
          title: Workflow Id
        workflow_name:
          type: string
          title: Workflow Name
        trigger_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Trigger Type
        status:
          type: string
          title: Status
        started_at:
          type: string
          format: date-time
          title: Started At
        finished_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Finished At
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
        requested_by:
          anyOf:
          - type: string
          - type: 'null'
          title: Requested By
      type: object
      required:
      - id
      - execution_id
      - workflow_id
      - workflow_name
      - status
      - started_at
      title: ExecutionHistoryEntry
      description: Single execution history row (cross-organization).
    ExecutionHistoryResponse:
      properties:
        entries:
          items:
            $ref: '#/components/schemas/ExecutionHistoryEntry'
          type: array
          title: Entries
        total:
          type: integer
          title: Total
      type: object
      required:
      - entries
      - total
      title: ExecutionHistoryResponse
      description: Paginated execution history across organizations.
    ExecutionStatisticsResponse:
      properties:
        window_hours:
          type: integer
          title: Window Hours
        counts_by_status:
          additionalProperties:
            type: integer
          type: object
          title: Counts By Status
        counts_by_workflow:
          additionalProperties:
            type: integer
          type: object
          title: Counts By Workflow
        counts_by_organization:
          additionalProperties:
            type: integer
          type: object
          title: Counts By Organization
        total_executions:
          type: integer
          title: Total Executions
      type: object
      required:
      - window_hours
      - counts_by_status
      - counts_by_workflow
      - counts_by_organization
      - total_executions
      title: ExecutionStatisticsResponse
      description: Execution statistics for time window analysis.
    ExecutionStatus:
      type: string
      enum:
      - queued
      - running
      - succeeded
      - failed
      - retrying
      - cancelled
      - timed_out
      title: ExecutionStatus
      description: Lifecycle of a workflow execution.
    ExecutionTimelineEvent:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        execution_id:
          type: string
          format: uuid
          title: Execution Id
        workflow_id:
          type: string
          format: uuid
          title: Workflow Id
        workflow_name:
          type: string
          title: Workflow Name
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        event_type:
          type: string
          title: Event Type
        status:
          type: string
          title: Status
        duration_ms:
          anyOf:
          - type: integer
          - type: 'null'
          title: Duration Ms
      type: object
      required:
      - id
      - execution_id
      - workflow_id
      - workflow_name
      - timestamp
      - event_type
      - status
      title: ExecutionTimelineEvent
      description: Single execution timeline event (cross-organization).
    ExecutionTimelineResponse:
      properties:
        events:
          items:
            $ref: '#/components/schemas/ExecutionTimelineEvent'
          type: array
          title: Events
      type: object
      required:
      - events
      title: ExecutionTimelineResponse
      description: Execution timeline across organizations.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    HeartbeatVisibilityResponse:
      properties:
        instance_id:
          type: string
          format: uuid
          title: Instance Id
        worker_type:
          type: string
          title: Worker Type
        pid:
          type: integer
          title: Pid
        hostname:
          type: string
          title: Hostname
        last_heartbeat_at:
          type: string
          format: date-time
          title: Last Heartbeat At
        loop_ok:
          type: boolean
          title: Loop Ok
        last_error:
          anyOf:
          - type: string
          - type: 'null'
          title: Last Error
        counters:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Counters
      type: object
      required:
      - instance_id
      - worker_type
      - pid
      - hostname
      - last_heartbeat_at
      - loop_ok
      title: HeartbeatVisibilityResponse
      description: Worker heartbeat visibility for monitoring.
...
    MonitoringInformationResponse:
      properties:
        system:
          $ref: '#/components/schemas/SystemHealth'
        database:
          $ref: '#/components/schemas/DatabaseHealth'
        workers:
          $ref: '#/components/schemas/WorkerHealthSummary'
        queue:
          $ref: '#/components/schemas/QueueMetrics'
      type: object
      required:
      - system
      - database
      - workers
      - queue
      title: MonitoringInformationResponse
      description: Comprehensive monitoring information payload.
...
    OperationalSummaryResponse:
      properties:
        timestamp:
          type: string
          format: date-time
          title: Timestamp
        executions:
          additionalProperties:
            type: integer
          type: object
          title: Executions
        events:
          additionalProperties:
            type: integer
          type: object
          title: Events
        workers:
          additionalProperties: true
          type: object
          title: Workers
        automation:
          additionalProperties: true
          type: object
          title: Automation
        schedules:
          additionalProperties:
            type: integer
          type: object
          title: Schedules
        system:
          additionalProperties:
            type: boolean
          type: object
          title: System
      type: object
      required:
      - timestamp
      - executions
      - events
      - workers
      - automation
      - schedules
      - system
      title: OperationalSummaryResponse
      description: Comprehensive operational summary for production dashboards.
...
    OrganizationQueue:
      properties:
        organization_id:
          type: string
          format: uuid
          title: Organization Id
        organization_name:
          type: string
          title: Organization Name
        queued_count:
          type: integer
          title: Queued Count
        running_count:
          type: integer
          title: Running Count
        pending_count:
          type: integer
          title: Pending Count
      type: object
      required:
      - organization_id
      - organization_name
      - queued_count
      - running_count
      - pending_count
      title: OrganizationQueue
      description: Per-organization queue metrics.
...
    QueueStatusResponse:
      properties:
        total_queued:
          type: integer
          title: Total Queued
        total_running:
          type: integer
          title: Total Running
        total_pending:
          type: integer
          title: Total Pending
        max_pending_per_org:
          type: integer
          title: Max Pending Per Org
        organization_queues:
          items:
            $ref: '#/components/schemas/OrganizationQueue'
          type: array
          title: Organization Queues
      type: object
      required:
      - total_queued
      - total_running
      - total_pending
      - max_pending_per_org
      - organization_queues
      title: QueueStatusResponse
      description: Aggregate queue status across all organizations.
...
    RetentionStatisticsResponse:
      properties:
        window_hours:
          type: integer
          title: Window Hours
        executions_deleted:
          type: integer
          title: Executions Deleted
        workers_pruned:
          type: integer
          title: Workers Pruned
      type: object
      required:
      - window_hours
      - executions_deleted
      - workers_pruned
      title: RetentionStatisticsResponse
      description: Retention worker statistics.
    ScheduleStatisticsResponse:
      properties:
        window_hours:
          type: integer
          title: Window Hours
        queued:
          type: integer
          title: Queued
        failed:
          type: integer
          title: Failed
        skipped:
          type: integer
          title: Skipped
        conflicts:
          type: integer
          title: Conflicts
      type: object
      required:
      - window_hours
      - queued
      - failed
      - skipped
      - conflicts
      title: ScheduleStatisticsResponse
      description: Schedule dispatch statistics.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
    WorkerStatisticsResponse:
      properties:
        window_hours:
          type: integer
          title: Window Hours
        active_workers:
          type: integer
          title: Active Workers
        healthy_loops:
          type: integer
          title: Healthy Loops
        errored_loops:
          type: integer
          title: Errored Loops
        health_percentage:
          type: number
          title: Health Percentage
        errors_by_type:
          additionalProperties:
            type: integer
          type: object
          title: Errors By Type
        workers:
          items:
            additionalProperties: true
            type: object
          type: array
          title: Workers
      type: object
      required:
      - window_hours
      - active_workers
      - healthy_loops
      - errored_loops
      - health_percentage
      - errors_by_type
      - workers
      title: WorkerStatisticsResponse
      description: Worker health and activity statistics.
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer